Add OtherFields to parser/generator
We need to be able to generate an anonymous

<definition>::Other

class with the same validations as ::Order and ::Invoice.

This isn't in use anywhere yet. It can only follow the
ContractFields and InvoiceFields, if it occurs at all, and
does not have a total_value_field.

We're going to wire it up to Ingest now.