Add 'other' fields to Ingest::*
Get the Loader and Converter to understand the concept of
"Other" fields (at the moment limited to the single safelist
worksheet name of "Briefs Received").

Add a spec in orchestrator_spec.rb as the outer integration.

At the moment loader_spec.rb has a mix of intra-spec variables
and let (for default_other_rows). We'll make this consistent
next.