Tidy up loader_spec.rb
Cut down on the manual `row_count`s in the specs.
While doing so, notice that that "happy path" spec
'loads data from the converter into the database'
has an issue with row counts – the new Other rows
were not being taken into consideration.

We noted that the Other rows were not being validated
correctly for a number of reasons:

1. There was no explicit inclusion of "other" in
   SubmissionEntry#entry_type; add a SubmissionEntry#TYPES
   constant and use that for inclusion
2. Our fake "other" row failed URN validation; set the
   URN to one that we've already created (12345).
3. Definition#for_entry_type didn't know about the 'Other'
   type and fell back to 'Order', which didn't exist. Add an
   explicit case statement for this (forgoing Hash or
   metaprogramming)