Loader should not store total_row_count
We noticed that the `rows` on a `SubmissionFile` was being updated in
both `Ingest::Loader` and `Ingest::Orchestrator`. It doesn't really
belong in loader, so we removed that.

This led to us wondering why `persist` was being passed in at all, so
we've removed it and the obsolete testing scripts that required it.