Don't error on entry types that aren't in FDL
Resolves https://rollbar.com/dxw/ccs-reportmi-api/items/133/ –
when users provided worksheets that contained entry_types that
weren't defined in the FDL, a `NameError` would be raised, e.g.

`NameError: uninitialized constant #<Class:0x00007fb5364131e0>::Order`

This prevents that from happening by asking the
anonymous `Framework::Definition` whether or not an `entry_type`
is defined before attempting to load entries for that type.

It also makes `total_row_count` aware of what the FDL has asked for
and no longer adds to the total for things that aren't defined
in the FDL. It takes the responsibility for counting from
`Ingest::Converter` and puts it in `Ingest::Orchestrator`, which
knows about the framework definition.

While we're about it remove the XLS fixture associated only with
that count.