Stop coercing URN to integer
The data warehouse export contained invalid URNs when a customer
submitted a value which contained a valid URN *followed by* additional
data (e.g. "12345 invalid").

This revealed two cases where Rails was unexpectedly calling `to_i` on
the URN:

1) In our anonymous ActiveModel-like class, we were declaring the URN
field as integer and `to_i` was biting us as the URN attribute was being
set via the intializer in `EntryData`
2) In URNValidator, `Customer.exists?(urn: value)`

This caused an unexpected behaviour which lead to invalid URNs being
incorrectly marked as valid, and subsequently an invalid submission
being submitted to CCS.

Stopping treating the URN as an integer in our anonymous class means no
coercion happens when setting the attribute.