Numeric fields no longer accept 'N/A'
Fields containing value equivalent of "not applicable" (e.g. 'N/A') have
been marked as valid for some time. This was left over from when were
onboarding the first few frameworks, and saw a large number of suppliers
submitting data containing N/A, instead of an appropriate value.

We now want to reject these values from numeric fields, and generate a
"is not a number" type validation error when this occurs.

The first step is to introduce a test case for Ingest::Loader to check
that we get the expected error. To get this passing,
`IngestedNumericalityValidator` is set to behave like ActiveModel's
standard `NumericalityValidator` instead.

Next, we'll remove the `IngestedNumericalityValidator` altogether.