Calculate column-based charges from other column
Allow column-based management charges to be optionally calculated from
another column. This is specified in FDL using the following form (the
'of' part is new):

```
ManagementCharge varies_by ‘a’, ‘b’ {
  ‘x’, ‘y’ -> 1% of ‘c’
  ‘x’, ‘z’ -> 2% of ‘d’
}
```

To do this we annotate the syntax tree differently. We have renamed
`:value` to `:percentage` in the parser. Because we've introduced the
concept of a `column`, percentages are always represented as hashes
rather than primitives. This means the transpiler has to pass a
different level of the hash to SectorBased management charges to
preserve the interface, whereas FlatRate only needed a rename.
ColumnBased deals with the percentage hashes directly.