Prevent exceptions when UserImportJob fails
Despite UserImportJob being configured to discard a job on an
`InvalidSaleforceId` exception, it kept retrying forever despite this
being a hard fail state.

Having looked at the production logs, it's possibly more to do with the
file being deleted from the S3 bucket whether or not the file was
processed successfully. I found at least one occasion where the call to
S3 to fetch the file failed (raising an Aws::S3::Errors::ServiceError)
exception, but the subsequent `delete_object` call passed, leaving the
job to retry with no file to process.

I've therefore reorganised the UserImportJob to only delete the file
after the importer finishes successfully. If the file legitimately does
not exist (due to S3 upload failure, for example) we additionally
discard the job when an `Aws::S3::Errors::NoSuchKey` exception occurs.

Finally, I'm only raising failures to Rollbar - the Rails logger is the
correct place for informational messages when the job succeeds.