Rename bulk supplier removal from lots
Because we're adding functionality to deactivate suppliers on Frameworks
in bulk, our previous approach of calling lot removal "Bulk Offboarding"
is not explicit enough. This is a large commit, but it's necessary to
rename all these classes/controllers in one go to ensure tests are all
green.

Changes:

- Rename Admin SupplierBulkOffboardsController -> SupplierBulkLotRemovalController
- Rename "offboarding" models to be more explicit about what they're actually doing
(removing suppliers from lots)
- Add new spec/fixture files to be more clear about what they're used
for